Updated !say command to prevent empty say commands

+++ src/index.js
@@ -23,13 +23,18 @@ client.on('message', msg => {
           "https://i.ibb.co/Zh8MY26/ED679-DCE-3-EEE-42-CE-8-D22-653-C5-D61-F268.jpg"
         ]})
         break;
-      default:
-        console.log(`Message by (id:${msg.author.id} Username: ${msg.author.username}#${msg.author.discriminator}): ${msg.content}`)
-        break;
+      // default:
+      //   console.log(`Message by (id:${msg.author.id} Username: ${msg.author.username}#${msg.author.discriminator}): ${msg.content}`)
+      //   break;
     }
     if(msg.content.substring(0,4) === '!say'){
-      msg.delete()
-      msg.channel.send(msg.content.substring(5))
+      if(msg.content.substring(5) != '') {
+        msg.delete()
+        msg.channel.send(msg.content.substring(5));
+      }
+      else{
+        msg.reply('Enter a message for me to repeat')
+      }
     }
 });
